feat(cdp): capture worker network events
