+++ api/src/services/cdp/instrumentation/page-events.ts
@@ -1,20 +1,11 @@
 import type { Page, CDPSession, TargetType, Protocol } from "puppeteer-core";
 import { BrowserEventType } from "../../../types/index.js";
 import { BrowserLogger } from "./browser-logger.js";
+import { attachNetworkEvents } from "./network-events.js";
+import type { AttachNetworkEventsOptions } from "./network-events.js";
 import { formatLocation, serializeRemoteObject } from "./utils.js";
 
-export interface AttachPageEventsOptions {
-  dangerouslyLogRequestDetails?: boolean;
-}
-
-const MAX_BODY_SIZE = 1_048_576; // 1 MB
-const TEXT_MIME_PREFIXES = ["text/", "application/json", "application/xml", "application/xhtml"];
-
-function isTextMime(mime: string | undefined): boolean {
-  if (!mime) return false;
-  const lower = mime.toLowerCase();
-  return TEXT_MIME_PREFIXES.some((p) => lower.startsWith(p));
-}
+export interface AttachPageEventsOptions extends AttachNetworkEventsOptions {}
 
 /**
  * Attach page-level event listeners. The caller must pass an already-enabled
@@ -29,7 +20,6 @@ export async function attachPageEvents(
   options?: AttachPageEventsOptions,
 ): Promise<void> {
   const pageId = (page.target() as any)._targetId as string;
-  const logBodies = options?.dangerouslyLogRequestDetails === true;
 
   // navigation
   page.on("framenavigated", (frame) => {
@@ -52,108 +42,9 @@ export async function attachPageEvents(
     navigation: { url: page.url() },
   });
 
-  // Track request metadata by requestId for use in loadingFailed (url) and loadingFinished (mimeType)
-  const requestMeta = new Map<string, { url: string; mimeType?: string }>();
-
-  // Network request logging via CDP Network domain.
-  // This fires for ALL requests including form POST navigations, unlike
-  // Puppeteer's page.on("request") which depends on Fetch interception
-  // and can miss requests during same-tab navigations.
-  session.on("Network.requestWillBeSent", (event: Protocol.Network.RequestWillBeSentEvent) => {
-    requestMeta.set(event.requestId, { url: event.request.url });
-
-    logger.record({
-      type: BrowserEventType.Request,
-      timestamp: new Date().toISOString(),
-      pageId,
-      targetType,
-      request: {
-        method: event.request.method,
-        url: event.request.url,
-        resourceType: event.type,
-        ...(logBodies && event.request.postData ? { postData: event.request.postData } : {}),
-        ...(logBodies && event.request.headers
-          ? { headers: event.request.headers as Record<string, string> }
-          : {}),
-      },
-    });
-  });
-
-  session.on("Network.responseReceived", (event: Protocol.Network.ResponseReceivedEvent) => {
-    const meta = requestMeta.get(event.requestId);
-    if (meta) {
-      meta.mimeType = event.response.mimeType;
-    }
-
-    const responseData: {
-      status: number;
-      url: string;
-      mimeType?: string;
-      headers?: Record<string, string>;
-    } = {
-      status: event.response.status,
-      url: event.response.url,
-      mimeType: event.response.mimeType,
-    };
-
-    if (logBodies && event.response.headers) {
-      responseData.headers = event.response.headers as Record<string, string>;
-    }
-
-    logger.record({
-      type: BrowserEventType.Response,
-      timestamp: new Date().toISOString(),
-      pageId,
-      targetType,
-      response: responseData,
-    });
-  });
-
-  // Always listen for loadingFinished to clean up requestMeta entries.
-  // When dangerouslyLogRequestDetails is enabled, also capture response bodies
-  // (size-capped, text-only MIME types).
-  session.on("Network.loadingFinished", (event: Protocol.Network.LoadingFinishedEvent) => {
-    const meta = requestMeta.get(event.requestId);
-    requestMeta.delete(event.requestId);
-
-    if (!logBodies) return;
-    if (event.encodedDataLength > MAX_BODY_SIZE) return;
-    if (!isTextMime(meta?.mimeType)) return;
-
-    session
-      .send("Network.getResponseBody", { requestId: event.requestId })
-      .then((result) => {
-        if (result?.body) {
-          logger.record({
-            type: BrowserEventType.ResponseBody,
-            timestamp: new Date().toISOString(),
-            pageId,
-            targetType,
-            responseBody: {
-              requestId: event.requestId,
-              body: result.body,
-              base64Encoded: result.base64Encoded,
-            },
-          });
-        }
-      })
-      .catch(() => {
-        // Response body not available (redirects, evicted, etc.) — ignore
-      });
-  });
-
-  session.on("Network.loadingFailed", (event: Protocol.Network.LoadingFailedEvent) => {
-    const url = requestMeta.get(event.requestId)?.url;
-    requestMeta.delete(event.requestId);
-
-    logger.record({
-      type: BrowserEventType.RequestFailed,
-      timestamp: new Date().toISOString(),
-      pageId,
-      targetType,
-      error: { message: event.errorText, url },
-    });
-  });
+  // This fires for ALL page-target requests including form POST navigations,
+  // unlike Puppeteer's page.on("request") which depends on Fetch interception.
+  attachNetworkEvents(session, logger, pageId, targetType, options);
 
   session.on("Runtime.consoleAPICalled", (event: Protocol.Runtime.ConsoleAPICalledEvent) => {
     const text = event.args.map(serializeRemoteObject).join(" ");
