+++ api/src/types/browser.ts
@@ -55,6 +55,7 @@ export interface BrowserLauncherOptions {
   deviceConfig?: { device: "desktop" | "mobile" };
   fullscreen?: boolean;
   dangerouslyLogRequestDetails?: boolean;
+  captureWorkerNetwork?: boolean;
   caCertificates?: string[];
 }
